restructure files & components
