+++ frontend/src/routes/_game.games.$gameId.play.galaxy.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router"
-import { GalaxyPage } from "@/features/play/GalaxyPage.tsx"
+import { GalaxyPage } from "@/features/play/galaxy/GalaxyPage.tsx"
 
 export const Route = createFileRoute("/_game/games/$gameId/play/galaxy")({
   component: GalaxyPage,
